feat: site overhaul — résumé, weather UOM, XR/metrics fixes, Docker Hub CI/CD
- Fix SVG images (hero logo, nav icon) via dangerouslyAllowSVG in next.config
- Regenerate all favicon PNGs from keycap SVG, replace old apple-touch-icon
- Résumé page: real career history (Cubic/Umo, Cubic Corp, Condé Nast),
  honest skill levels, proper diacritics, TKS entry removed
- Blog page: tone down aspirational framing, honest empty state
- Add 30 switches via V8 migration (total 50 across 18+ manufacturers)
- Weather: default to US Imperial, persist UOM toggle in localStorage,
  reverse geocode coordinates to show actual city/state via Nominatim
- XR: add blob: to script-src and raw.githubusercontent.com to connect-src
  in CSP, add Inter Bold font for 3D text rendering
- Metrics: cast Prometheus doubles to long for GraphQL Long! fields
- Light/dark toggle: add .light class, use :root:not(.light) for dark
  media query so manual toggle overrides OS preference
- Docker Hub CI/CD: 3-stage API Dockerfile with layered JAR + non-root,
  multi-stage frontend Dockerfile with non-root, docker-compose.yml uses
  image: refs, GitHub Actions builds/pushes to Docker Hub then deploys by pull
- Update ARCHITECTURE.md for registry-backed deploy pattern
- Add DOCKER_HUB_INTEGRATION.md operator/agent guide
- Fix Nav test for Résumé label change

+++ api/src/main/java/com/thekeyswitch/api/service/MetricsProxyService.java
@@ -20,11 +20,11 @@ public class MetricsProxyService {
         Map<String, Object> metrics = new HashMap<>();
 
         metrics.put("cpuUsagePercent", queryPrometheus("100 - (avg(rate(node_cpu_seconds_total{mode=\"idle\"}[1m])) * 100)"));
-        metrics.put("memoryUsedBytes", queryPrometheus("node_memory_MemTotal_bytes - node_memory_MemAvailable_bytes"));
-        metrics.put("memoryTotalBytes", queryPrometheus("node_memory_MemTotal_bytes"));
-        metrics.put("diskUsedBytes", queryPrometheus("node_filesystem_size_bytes{mountpoint=\"/\"} - node_filesystem_avail_bytes{mountpoint=\"/\"}"));
-        metrics.put("diskTotalBytes", queryPrometheus("node_filesystem_size_bytes{mountpoint=\"/\"}"));
-        metrics.put("uptimeSeconds", queryPrometheus("node_time_seconds - node_boot_time_seconds"));
+        metrics.put("memoryUsedBytes", (long) queryPrometheus("node_memory_MemTotal_bytes - node_memory_MemAvailable_bytes"));
+        metrics.put("memoryTotalBytes", (long) queryPrometheus("node_memory_MemTotal_bytes"));
+        metrics.put("diskUsedBytes", (long) queryPrometheus("node_filesystem_size_bytes{mountpoint=\"/\"} - node_filesystem_avail_bytes{mountpoint=\"/\"}"));
+        metrics.put("diskTotalBytes", (long) queryPrometheus("node_filesystem_size_bytes{mountpoint=\"/\"}"));
+        metrics.put("uptimeSeconds", (long) queryPrometheus("node_time_seconds - node_boot_time_seconds"));
         metrics.put("loadAverage1m", queryPrometheus("node_load1"));
         metrics.put("loadAverage5m", queryPrometheus("node_load5"));
         metrics.put("loadAverage15m", queryPrometheus("node_load15"));
